+++ script/strategies/mangroveOrder/deployers/MumbaiMangroveOrderDeployer.s.sol
@@ -5,7 +5,7 @@ import {Script, console} from "@mgv/forge-std/Script.sol";
 import {MangroveOrder, IERC20, IMangrove} from "@mgv-strats/src/strategies/MangroveOrder.sol";
 
 import {Deployer} from "@mgv/script/lib/Deployer.sol";
-import {MangroveOrderDeployer} from "./MangroveOrderDeployer.s.sol";
+import {MangroveOrderDeployer, RouterProxyFactory} from "./MangroveOrderDeployer.s.sol";
 
 /**
  * Mumbai specific deployment of MangroveOrderDeployer
@@ -19,6 +19,7 @@ contract MumbaiMangroveOrderDeployer is Deployer {
   function runWithChainSpecificParams() public {
     new MangroveOrderDeployer().innerRun({
       mgv: IMangrove(envAddressOrName("MGV", "Mangrove")),
+      routerProxyFactory: RouterProxyFactory(envAddressOrName("ROUTER_PROXY_FACTORY", "RouterProxyFactory")),
       admin: envAddressOrName("MGV_GOVERNANCE", broadcaster())
     });
   }