+++ src/strategies/routers/abstract/AbstractRouter.sol
@@ -3,14 +3,12 @@ pragma solidity ^0.8.10;
 
 import {AccessControlled} from "@mgv-strats/src/strategies/utils/AccessControlled.sol";
 import {IERC20} from "@mgv/lib/IERC20.sol";
+import {RoutingOrderLib as RL} from "./RoutingOrderLib.sol";
 
 /// @title AbstractRouter
 /// @notice Partial implementation and requirements for liquidity routers.
 
 abstract contract AbstractRouter is AccessControlled(msg.sender) {
-  ///@notice the bound maker contracts which are allowed to call this router.
-  mapping(address => bool) internal boundMakerContracts;
-
   ///@notice This modifier verifies that `msg.sender` an allowed caller of this router.
   modifier onlyBound() {
     require(isBound(msg.sender), "AccessControlled/Invalid");
@@ -36,58 +34,58 @@ abstract contract AbstractRouter is AccessControlled(msg.sender) {
   ///@notice getter for the `makers: addr => bool` mapping
   ///@param mkr the address of a maker contract
   ///@return true if `mkr` is authorized to call this router.
-  function isBound(address mkr) public view returns (bool) {
-    return boundMakerContracts[mkr];
+  function isBound(address mkr) public view virtual returns (bool) {
+    return RL.boundMakerContracts()[mkr];
   }
 
   ///@notice pulls liquidity from the reserve and sends it to the calling maker contract.
-  ///@param token is the ERC20 managing the pulled asset
-  ///@param reserveId identifies the fund owner (router implementation dependent).
-  ///@param amount of `token` the maker contract wishes to pull from its reserve
-  ///@param strict when the calling maker contract accepts to receive more funds from reserve than required (this may happen for gas optimization)
-  ///@return pulled the amount that was successfully pulled.
-  function pull(IERC20 token, address reserveId, uint amount, bool strict) external onlyBound returns (uint pulled) {
+  ///@param routingOrder the arguments of the pull order
+  ///@param amount of token that needs to be routed
+  ///@param strict if false the router may pull at more than `amount` to msg.sender. Otherwise it pulls at least `amount`.
+  ///@return pulled the amount of `routingOrder.token` that has been sent to `msg.sender`
+  function pull(RL.RoutingOrder calldata routingOrder, uint amount, bool strict)
+    external
+    onlyBound
+    returns (uint pulled)
+  {
     if (strict && amount == 0) {
       return 0;
     }
-    pulled = __pull__({token: token, reserveId: reserveId, amount: amount, strict: strict});
+    pulled = __pull__(routingOrder, amount, strict);
   }
 
-  ///@notice router-dependent implementation of the `pull` function
-  ///@param token Token to be transferred
-  ///@param reserveId determines the location of the reserve (router implementation dependent).
-  ///@param amount The amount of tokens to be transferred
-  ///@param strict wether the caller maker contract wishes to pull at most `amount` tokens of owner.
-  ///@return pulled The amount pulled if successful; otherwise, 0.
-  function __pull__(IERC20 token, address reserveId, uint amount, bool strict) internal virtual returns (uint);
+  ///@notice router dependent hook to customize pull orders.
+  ///@param routingOrder the arguments of the pull order
+  ///@param amount of token that needs to be routed
+  ///@param strict if false the router may pull at more than `amount` to msg.sender. Otherwise it pulls at least `amount`.
+  ///@return pulled the amount of `routingOrder.token` that has been sent to `msg.sender`
+  function __pull__(RL.RoutingOrder memory routingOrder, uint amount, bool strict) internal virtual returns (uint);
 
-  ///@notice pushes assets from calling's maker contract to a reserve
-  ///@param token is the asset the maker is pushing
-  ///@param reserveId determines the location of the reserve (router implementation dependent).
-  ///@param amount is the amount of asset that should be transferred from the calling maker contract
-  ///@return pushed fraction of `amount` that was successfully pushed to reserve.
-  function push(IERC20 token, address reserveId, uint amount) external onlyBound returns (uint pushed) {
+  ///@notice pushes liquidity from msg.sender to the reserve
+  ///@param routingOrder the arguments of the push order
+  ///@param amount of token that needs to be routed
+  ///@return pushed the amount of `routingOrder.token` that has been taken from `msg.sender`
+  function push(RL.RoutingOrder calldata routingOrder, uint amount) external onlyBound returns (uint pushed) {
     if (amount == 0) {
       return 0;
     }
-    pushed = __push__({token: token, reserveId: reserveId, amount: amount});
+    pushed = __push__(routingOrder, amount);
   }
 
-  ///@notice router-dependent implementation of the `push` function
-  ///@param token Token to be transferred
-  ///@param reserveId determines the location of the reserve (router implementation dependent).
-  ///@param amount The amount of tokens to be transferred
-  ///@return pushed The amount pushed if successful; otherwise, 0.
-  function __push__(IERC20 token, address reserveId, uint amount) internal virtual returns (uint pushed);
+  ///@notice router dependent hook to customize pull orders.
+  ///@param routingOrder the arguments of the pull order
+  ///@param amount of token that needs to be routed
+  ///@return pushed the amount of `routingOrder.token` that has been sent to `msg.sender`
+  function __push__(RL.RoutingOrder memory routingOrder, uint amount) internal virtual returns (uint pushed);
 
-  ///@notice iterative `push` for the whole balance in a single call
-  ///@param tokens to flush
-  ///@param reserveId determines the location of the reserve (router implementation dependent).
-  function flush(IERC20[] calldata tokens, address reserveId) external onlyBound {
-    for (uint i = 0; i < tokens.length; ++i) {
-      uint amount = tokens[i].balanceOf(msg.sender);
+  ///@notice iterative `push` of the whole maker contract's balance
+  ///@dev minimizes external calls in case several assets needs to be pushed via the router.
+  ///@param routingOrders to be executed
+  function flush(RL.RoutingOrder[] memory routingOrders) external onlyBound {
+    for (uint i = 0; i < routingOrders.length; ++i) {
+      uint amount = routingOrders[i].token.balanceOf(msg.sender);
       if (amount > 0) {
-        require(__push__(tokens[i], reserveId, amount) == amount, "router/pushFailed");
+        require(__push__(routingOrders[i], amount) == amount, "router/flushFailed");
       }
     }
   }
@@ -96,14 +94,14 @@ abstract contract AbstractRouter is AccessControlled(msg.sender) {
   ///@dev this function is callable by router's admin to bootstrap, but later on an allowed maker contract can add another address
   ///@param makerContract the maker contract address
   function bind(address makerContract) public onlyAdmin {
-    boundMakerContracts[makerContract] = true;
+    RL.boundMakerContracts()[makerContract] = true;
     emit MakerBind(makerContract);
   }
 
   ///@notice removes a maker contract address from the allowed makers of this router
   ///@param makerContract the maker contract address
   function _unbind(address makerContract) internal {
-    boundMakerContracts[makerContract] = false;
+    RL.boundMakerContracts()[makerContract] = false;
     emit MakerUnbind(makerContract);
   }
 
@@ -118,38 +116,10 @@ abstract contract AbstractRouter is AccessControlled(msg.sender) {
     _unbind(makerContract);
   }
 
-  ///@notice allows a makerContract to verify it is ready to use `this` router for a particular reserve
-  ///@dev `checkList` returns normally if all needed approval are strictly positive. It reverts otherwise with a reason.
-  ///@param token is the asset (and possibly its overlyings) whose approval must be checked
-  ///@param reserveId of the tokens that are being pulled
-  function checkList(IERC20 token, address reserveId) external view {
-    require(isBound(msg.sender), "Router/callerIsNotBoundToRouter");
-    // checking maker contract has approved this for token transfer (in order to push to reserve)
-    require(token.allowance(msg.sender, address(this)) > 0, "Router/NotApprovedByMakerContract");
-    // pulling on behalf of `reserveId` might require a special approval (e.g if `reserveId` is some account on a protocol).
-    __checkList__(token, reserveId);
-  }
-
-  ///@notice router-dependent additional checks
-  ///@param token is the asset (and possibly its overlyings) whose approval must be checked
-  ///@param reserveId of the tokens that are being pulled
-  function __checkList__(IERC20 token, address reserveId) internal view virtual;
-
-  ///@notice performs necessary approval to activate router function on a particular asset
-  ///@param token the asset one wishes to use the router for
-  function activate(IERC20 token) external boundOrAdmin {
-    __activate__(token);
-  }
-
-  ///@notice router-dependent implementation of the `activate` function
-  ///@param token the asset one wishes to use the router for
-  function __activate__(IERC20 token) internal virtual {
-    token; //ssh
-  }
-
-  ///@notice Balance of a reserve
-  ///@param token the asset one wishes to know the balance of
-  ///@param reserveId the identifier of the reserve
-  ///@return the balance of the reserve
-  function balanceOfReserve(IERC20 token, address reserveId) public view virtual returns (uint);
+  /**
+   * @notice Returns the token balance available for a specific routing order
+   * @param routingOrder The routing order to check the balance for
+   * @return balance The balance of the token in the routing order
+   */
+  function tokenBalanceOf(RL.RoutingOrder calldata routingOrder) public view virtual returns (uint balance);
 }