+++ CHANGELOG.md
@@ -5,6 +5,41 @@
 - Upgrade to @mangrovedao/context-addresses v1.1.0
 - Simplify copying of context addresses
 
+# 2.0.0-b1.3
+
+- Add export for `MangroveAmplifier`
+- Add deployer for `SimpleAaveLogic`
+- Add AavePooledRouter to ABI exports
+- Upgrade to @mangrovedao/mangrove-deployments v2.0.1-1
+
+# 2.0.0-b1.2
+
+- Add exports for `AbstractRoutingLogic` and `SimpleAaveLogic`
+
+# 2.0.0-b1.1
+
+- Upgrade to @mangrovedao/mangrove-deployments v2.0.0
+- Add Mock Aave Oracle to Mock Aave contract
+
+# 2.0.0-b1.0
+
+- feat!: Remove mangrove.js deployer
+- Created 'SmartRouter' schema to allow to plug any RoutingLogic to a ForwarderContract
+- Created 'RouterProxyFactory' to deploy router proxies for each user and allow all approvals to a single contract per user
+- Added 'RenegingForwarder' to allow time based and volume based reneging
+- Changed 'IOfferLogic' 'router' view function to be passed an argument, supposedly the 'owner' of the router
+- Added a 'router' view function with no args to 'Direct' contract because Direct only needs a single router
+- Changed MangroveOrder
+  - Default Router is now a proxied Smart Router
+  - 'take' entrypoints args changed
+    - Removed 'fillOrKill' and 'restingOrder' boolean args
+    - Added 'takerOrderType' enum with GTC(E), IOC, FOK, and PO orders
+  - MangroveOrder now extends RenegingForwarder
+- Added 'MangroveAmplifier' to amplify create offers on multiple market from a single token
+  - Possibility to attach a logic to every token
+  - Bundle-wise reneging on time possibility
+- Added Aave Routing logic
+
 # 1.0.2
 
 - Add support for Sepolia